adding new version and normalized version compute rp
