+++ compute_rp/20181001/module.ts
@@ -12,7 +12,7 @@ import * as armtypes from "@azure-tools/arm.adl"; // arm extentions to adl
 // As an rp owner, these are my stuff
 
 // Importing my version-ed types
-import { VirtualMachineScaleSet20180601 } from "./vmscalesets";
+import { VirtualMachineScaleSet20181001 } from "./vmscalesets";
 
 // importing my normalized types
 import * as core from "../normalized/module";
@@ -20,11 +20,11 @@ import * as core from "../normalized/module";
 // vm scale set is a top level resource
 // note: It does not require any special conversion logic so wre are composing it with  AutoVersioned<T>
 // Auto converter can convert between versions based on field names.
-export type vmscaleset_resource_20180601 = adltypes.ApiType<
+export type vmscaleset_resource_20181001 = adltypes.ApiType<
   "vmscalesetnormalized",
-  "vmscaleset20180601",
+  "vmscaleset20181001",
   core.VirtualMachineScaleSetNormalized,
-  VirtualMachineScaleSet20180601
+  VirtualMachineScaleSet20181001
 >;
 
 // any imperative code needs to be exported at the package level
