+++ compute_rp/normalized/module.ts
@@ -6,6 +6,7 @@ import { VirtualMachineNormalized } from "./vm";
 import { ResourceTwoNormalized, RespourceTwoNormalizer } from "./resource_two";
 import { ResourceThreeProps } from "./resource_three";
 import { ResourceFourProps } from "./resource_four";
+import { VirtualMachineScaleSetNormalized } from "./vmscalesets";
 
 // normalized resources
 
@@ -36,11 +37,18 @@ export type ResourceFour = adltypes.NormalizedApiType<
   ResourceFourProps
 >;
 
+export type VirtualMachineScaleSet = adltypes.NormalizedApiType<
+  "vmscalesetnormalized",
+  VirtualMachineScaleSetNormalized
+>;
+
+
 // export properties so versioned decleration can see it
 export * from "./vm";
 export * from "./resource_two";
 export * from "./resource_three";
 export * from "./resource_four";
+export * from "./vmscalesets";
 
 // TODO: remove when loadable runtime feat is done
 // we need that because the load logic verify that the library loaded has the
